requirements.txt
pytest
pyside6
pyyaml
nuitka
pytest-qt
ruamel.yaml
